Allow multiple streaming operations to be started
When trying to subscribe to multiple UserStream topics, only the
first one was succeeding (since it got paused and then resumed on
didReceiveMessage in the isInitialConnectionResponse state). This
pauses subsequent operations that come in while we're trying to
connect the websocket so once we successfully connect, we can
resume (and subscribe with the correct connection_id).